Clip Confirm 1.0.0
Initial release of a new add-on. Clip Confirm adds a confirmation window when the Clip button is clicked/pressed in an attempt to help prevent accidental clips. Also allows for a "Skip Confirmation" bypass hotkey to allow for manually skipping the confirmation window in the event you 100% know you mean to be making a clip.